moved booking form in to booking field
