+++ imports/ui/pages/resources/components/BookingsField.jsx
@@ -1,21 +1,26 @@
 import { Meteor } from 'meteor/meteor';
 import React, { useState, useEffect } from 'react';
+import { useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
-import { Box,  Heading, Text, Button } from '@chakra-ui/react';
+import { Box, Flex, Heading, Text, Button, HStack, Textarea } from '@chakra-ui/react';
 import moment from 'moment';
 
 import { call } from '../../../@/shared';
 import NiceList from '../../../components/NiceList';
 import { message } from '../../../components/message';
-import BookingForm from './BookingForm';
+import DatePicker from '../../../components/DatePicker';
 
 export default function BookingsField({ domainId }) {
   const [ t ] = useTranslation('processes');
   const [ tc ] = useTranslation('common');
   
   const [ bookings, setBookings] = useState([]);
+  const [ newBooking, setNewBooking ] = useState({});
   const [ isLoading, setIsLoading ] = useState(true);
   const [ isAdmin ] = useState(true);
+
+  const { formState, handleSubmit, register } = useForm();
+  const { isDirty, isSubmitting } = formState;
 
   useEffect(() => {
     getBookings();
@@ -24,7 +29,7 @@ export default function BookingsField({ domainId }) {
 
   const bookingListActions = [{
     content: tc('labels.remove'),
-    // handleClick: () => removeDocument(booking._id),
+    // handleClick: () => removeBooking(booking._id),
   }];
 
   const getBookings = async () => {
@@ -39,7 +44,27 @@ export default function BookingsField({ domainId }) {
     }
   };
     
-    
+  const handleDateAndTimeChange = (value, key) => {
+    newBooking[key] = value;
+    setNewBooking({...newBooking});
+  };
+
+  const onSubmit = async (values) => {
+    values = { ...newBooking, endDate: newBooking.startDate, ...values }
+    Meteor.call(
+      'createBooking',
+      domainId,
+      values,
+      (error, respond) => {
+        if (error) {
+          console.log('error', error);
+          message.error(error.error);
+        } else {
+          message.success(t('meeting.success.add'));
+        }
+      }
+    );
+  };
 
   return (
     <Box mt="1.75rem">
@@ -47,7 +72,49 @@ export default function BookingsField({ domainId }) {
         Bookings
       </Heading>
 
-      {isAdmin && <BookingForm domainId={domainId} />}
+      {isAdmin &&       
+        <Box px="2" py="4" bg="white">
+          <form onSubmit={handleSubmit((data) => onSubmit(data))}>
+            <Box mb="4">
+              <DatePicker 
+                noTime 
+                onChange={(date) => handleDateAndTimeChange(date, 'startDate')} 
+              />
+            </Box>
+            <HStack spacing="2" mb="4">
+              <DatePicker
+                onlyTime
+                placeholder={t('meeting.form.time.start')}
+                onChange={(time) => handleDateAndTimeChange(time, 'startTime')}
+              />
+              <DatePicker
+                onlyTime
+                placeholder={t('meeting.form.time.end')}
+                onChange={(time) => handleDateAndTimeChange(time, 'endTime')}
+              />
+            </HStack>
+
+            <Textarea
+              {...register('description')}
+              placeholder="Note, usage, purpose, etc..."
+              size="sm"
+              mb="4"
+            />
+
+            <Flex justify="flex-end">
+              <Button
+                colorScheme="green"
+                isDisabled={!isDirty}
+                isLoading={isSubmitting}
+                size="sm"
+                type="submit"
+              >
+                Book
+              </Button>
+            </Flex>
+          </form>
+        </Box>
+      }
 
       {!isLoading && 
         <Box
